requirements.txt
wandb
tqdm
matplotlib
lightning
torchdyn
ray
